Add support for return values in Computation #40
Does not include tests yet.
- added representation of return values to Computation
- changed appendComputation to return a new struct which includes previous
  tuple position int, as well as return values
- allowed literal arguments to appendComputation, treated similarly to literal
  return values
- renamed 'level' parameter of appendComputation to 'depth' for clarity

This code rests on the assumption that if an argument name/return value name
cannot be found in the list of data spaces of its Computation, it is a literal
of the appropriate type. I believe this assumption is sound, as anything else
would correspond to a variable in code that was not accounted for in the data
spaces (indicating a problem elsewhere).