Increase uniqueness of cloned Computation name prefixes
Each appended instance of a Computation now gets a universally unique prefix
instead of just unique for that Computation. This is done by making the rename
counter static. This is needed because previously, different Computations
appended in the same context that have some identical names could have name
collisions.

Tests are adjusted to account for this change. Test SetUp() now includes
resetting the counter between tests, and to take proper advantage of
this the numerous different "sub-tests" in the test "AppendComputation"
have been split apart into separate tests.