Improve appendComputation testing
Adds an 'expectComputationsEqual' helper function that gtest EXPECTs,
component by component, that two Computations are the same. This
function is called in checkAppendComputation, so now you have to pass
in what you expect the appender to look like after appending.
Some tests currently fail, which is expected now that we are testing the
entire append result instead of just implemented pieces. This holistic
test is better now that appendComputation is almost complete. Existing
tests will definitely need to be augmented to give more complex input
and expect more complex output, to test handling of things like complex
iteration domains.