Completed Unit tests for OmegaVisitor to convert from iegenlib to omega compliant ufs in the presence of nesting.
Added tests for computation object using this extension.

Test passing for both cases and tackles issue #2

Started work on infinite level of nesting, tests currently
failing.

Note: possible execution schedule bug

Deciding where to locate the tuple variable for nesting is done
by tuple shuffling, where we place new tuple variable for the
nested uf as close as possible to the maximum tuple variable present in that
expression

A major challenge is tuple shuffle could possibly mess up the execution schedule.
This is because the call to make a set omega compliant is made after the
execution schedule is applied to the set in the computation.